refactoring of giskard_evaluator
